Extract EscapeYamlSpecialCharacters
